Update to only accept a single EC2 instance ID vice a list of them
Terraform is unable to apply changes to this module for a single EC2
instance in cisagov/cool-sharedservices-freeipa where, when a new
FreeIPA AMI is available, the full list of instance IDs cannot be
constructed prior to redeploying all FreeIPA instances.

With this change the DNS records with health checks associated with
the FreeIPA cluster should now be able to function as expected, which
should in turn enable us to upgrade the FreeIPA cluster with zero
downtime.